cmt 33 [30/08]
Se agrega autenticador el cual recibirá el message context previamente
tomado del servicio mediante request context, y verificará los headers
en busqueda de usr y pass. dependiendo de estos datos, permitirá o
denegará la ejecucion de la operacion del servicio. Resta agregar (prop
o bd), el ust y pass del lado del servidor para poder controlar contra
lo que viene desde afuera (por ahora se dejan datos soldados).

+++ GestPedWebDataAccess/src/gpw/dominio/producto/Producto.java
@@ -15,7 +15,7 @@ public class Producto implements Serializable {
 	private String nombre;
 	private String descripcion;
 	private Unidad unidad;
-	private Integer cantUnidad;
+	private Float cantUnidad;
 	private AplicaIva aplIva;
 	private Double precioVta;
 	private Sinc sinc;
@@ -65,10 +65,10 @@ public class Producto implements Serializable {
 		this.unidad = unidad;
 	}
 	
-	public Integer getCantUnidad() {
+	public Float getCantUnidad() {
 		return cantUnidad;
 	}
-	public void setCantUnidad(Integer cantUnidad) {
+	public void setCantUnidad(Float cantUnidad) {
 		this.cantUnidad = cantUnidad;
 	}
