+++ GestPedWebDataAccess/src/gpw/persistencia/producto/PersistenciaProducto.java
@@ -50,7 +50,7 @@ public class PersistenciaProducto extends Conector implements IPersProducto, Cns
 					AplicaIva aplIva = AplicaIva.getAplicaIvaPorChar(aplIvaChar[0]);
 					producto.setAplIva(aplIva);
 					producto.setUnidad(pu.obtenerUnidadPorId(conn, rs.getInt("id_unidad")));
-					producto.setCantUnidad(rs.getInt("cant_unidad"));
+					producto.setCantUnidad(rs.getFloat("cant_unidad"));
 					producto.setPrecioVta(rs.getDouble("precio_vta"));
 					char[] sincChar = new char[1];
 					rs.getCharacterStream("sinc").read(sincChar);
@@ -91,7 +91,7 @@ public class PersistenciaProducto extends Conector implements IPersProducto, Cns
 					AplicaIva aplIva = AplicaIva.getAplicaIvaPorChar(aplIvaChar[0]);
 					producto.setAplIva(aplIva);
 					producto.setUnidad(pu.obtenerUnidadPorId(conn, rs.getInt("id_unidad")));
-					producto.setCantUnidad(rs.getInt("cant_unidad"));
+					producto.setCantUnidad(rs.getFloat("cant_unidad"));
 					producto.setPrecioVta(rs.getDouble("precio_vta"));
 					char[] sincChar = new char[1];
 					rs.getCharacterStream("sinc").read(sincChar);
